# Optimize Active Record batching for whole table iterations