@fatkodima 2022-06-21
Added test cases.