@matthewd 2022-06-21
Given it's (so measurably) fast enough to matter, it does feel to me like we should be exposing this as a [rarely used, but available] option to override this heuristic... if the only condition in play is a `archived_at: nil` (and only a tiny fraction of your records are archived), it may well still make sense to opt in to this approach?