@fatkodima 2022-06-21
> it may well still make sense to opt in to this approach

I think yes, we just need a good (highlevel?) name for this option 😅