@simi 2022-06-21
I'll test this on real data and report back. Btw. @fatkodima it is good practice on PostgreSQL after each big insert of data to run ANALYZE on table to get proper statistics for query planner.