@simi 2022-06-21
> > it is good practice on PostgreSQL after each big insert of data to run ANALYZE on table to get proper statistics for query planner
> 
> Sure, but we index only scan the primary key index here, so this should not make a difference.

```sql
DROP TABLE users1;
CREATE TABLE users1 (id bigserial PRIMARY KEY, val integer DEFAULT 0);
INSERT INTO users1 SELECT i FROM generate_series(1, 10000000) AS i;
EXPLAIN SELECT id FROM users1 WHERE id IN (1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
ANALYZE users1;
EXPLAIN SELECT id FROM users1 WHERE id IN (1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
```

:thinking: I get different costs for both queries. I'm not sure how that is affecting your numbers, but I was just about to mention it is good idea to start with analyzed data.