@fatkodima 2022-06-27
@simi Were you able to test this on real data - https://github.com/rails/rails/pull/45414#issuecomment-1161590945 ?