@simi 2022-06-28
> > I can check again later on EXPLAIN ANALYZE at least in transaction and rollback on my testing data to get better insight.
> 
> +1 Even though the query plan may look scarier, the real execution time can be better.
> 
> It would be great, if you can suggest some sort of a reproduction script (to be able to test locally), that proves the new approach can be slower.

I'm trying to point to the problem that `delete_all` on batch is being transformed in different way to `DELETE` query (using subquery) with this change. Good next step would be to check `update_all`.

> This list is often very large (1000 by default), which can lead to inefficient queries and truncated sql queries in logs and monitoring tools. I'm sure everyone saw that monstrous queries of the half of the screen in size smile

This was the original description and motivation for this change. Maybe there is actual reason for the current implementation. I can try to get more testing results if welcomed. :pray: