@simi 2022-06-28
> > I'm trying to point to the problem that delete_all on batch is being transformed in different way to DELETE query (using subquery) with this change. Good next step would be to check update_all.
> 
> I measured `delete_all/update_all` (see PR description), but yes, I used only the default batch size of 1000. I will try to test with bigger tables and different batch sizes.

I'm related to PostgreSQL only. I'm able to evaluate only query plans. Would you mind to share query plans as well for your testing in PR description?