@fatkodima 2022-06-29
Updated description with query plans.

Ok, tested with the same 10m table and 100k batch size this time.

Select: `238s` vs `7s`. 🔥 🔥 🔥 
Update: `204s` vs `104s`. 🔥 
Delete: `172s` vs `323s` 😱.

So the new approach can be slower.

But 100k is kinda unrealistic batch size, from my experience.

Tested deletes with other batch sizes:
* 10k: `114s` vs `69s`
* 30k: `129s` vs `69s`
* 50k: `136s` vs `61s`