@fatkodima 2022-07-04
@byroot 
1. Wdyt about Matthew's suggestion at https://github.com/rails/rails/pull/45414#issuecomment-1161489698?

2. And given this PR can make (while very unlikely) the batching slower (see the last comment), should this PR's optimization be applied by default or must be enabled by the option from point 1?