@byroot 2022-07-05
> 1. Wdyt about Matthew's suggestion at

Yeah, I think having a param to manually opt-in to this bahavior on arbitrary scope would make perfect sense.

> should this PR's optimization be applied by default 

I think it should be used by default for empty scopes yes.