@byroot 2022-07-05
> Have you checked my finding

Yes, and unless I missed something, it's only less performance on really huge batch sizes, so I'm not too worried.